Extract broker presentation from AgentBrokerService
Peel session-open handoff assembly, remember payload/supersede
selection, and recall hit wire into package enums. Move markdown
export helpers onto the existing +Markdown extension. handle(_:)
stays the caller-facing seam; wire keys unchanged.
